+++ src/m3p2i_aip/config/isaacgym/point.yaml
