+++ src/m3p2i_aip/config/panda_env/3_shelf_stand.yaml
