+++ src/m3p2i_aip/config/point_env/0_point_robot.yaml
