+++ src/m3p2i_aip/config/point_env/6_dyn_obs.yaml
