+++ src/m3p2i_aip/config/point_env/7_box.yaml
